refactor: extract chart components and analytics utils from large page files
