+++ frontend/src/features/courses/pages/CourseDetail.tsx
@@ -20,34 +20,18 @@ import { coreStudentLoScoresList } from '../../../shared/api/generated/scores/sc
 import { evaluationGradesList, evaluationEnrollmentsList, evaluationAssessmentsDestroy } from '../../../shared/api/generated/evaluation/evaluation'
 import { Card } from '../../../shared/components/ui/Card'
 import { ChartWidget } from '../../../shared/components/ui/ChartWidget'
-import {
-  BookOpenIcon,
-  UsersIcon,
-  AcademicCapIcon,
-  ChartBarIcon,
-} from '@heroicons/react/24/outline'
+import { CourseHeader } from '../components/CourseHeader'
+import { BoxPlotChart, BoxPlotLegend } from '../components/BoxPlotChart'
+import { StudentHeatmap, AssessmentHeatmap } from '../components/StudentHeatmap'
+import { LoRadarChart } from '../components/LoRadarChart'
+import type { BoxPlotData } from '../components/BoxPlotChart'
+import type { HeatmapData } from '../components/StudentHeatmap'
 import type {
   Course,
   CourseInstructorsItem,
   CoreLearningOutcome,
   StudentLearningOutcomeScore,
 } from '../../../shared/api/model'
-
-interface BoxPlotData {
-  code: string
-  min: number
-  q1: number
-  median: number
-  q3: number
-  max: number
-  mean: number
-}
-
-interface HeatmapData {
-  studentName: string
-  studentId: number
-  loScores: Record<string, number>
-}
 
 interface AssessmentHeatmapData {
   assessments: { id: number; name: string; totalScore: number }[]
@@ -474,44 +458,6 @@ const CourseDetail = () => {
     })
   }, [gradesData])
 
-  const getHeatmapColor = (score: number): string => {
-    if (score === 0) return 'rgb(249, 250, 251)'
-
-    const normalized = Math.max(0, Math.min(100, score)) / 100
-
-    if (normalized < 0.25) {
-      const t = normalized / 0.25
-      const r = 239 - Math.round(11 * t)
-      const g = 68 + Math.round(46 * t)
-      const b = 68 + Math.round(14 * t)
-      return `rgb(${r}, ${g}, ${b})`
-    } else if (normalized < 0.5) {
-      const t = (normalized - 0.25) / 0.25
-      const r = 228 + Math.round(24 * t)
-      const g = 114 + Math.round(97 * t)
-      const b = 82 - Math.round(27 * t)
-      return `rgb(${r}, ${g}, ${b})`
-    } else if (normalized < 0.75) {
-      const t = (normalized - 0.5) / 0.25
-      const r = 252 - Math.round(68 * t)
-      const g = 211 + Math.round(33 * t)
-      const b = 55 + Math.round(58 * t)
-      return `rgb(${r}, ${g}, ${b})`
-    } else {
-      const t = (normalized - 0.75) / 0.25
-      const r = 184 - Math.round(150 * t)
-      const g = 244 - Math.round(42 * t)
-      const b = 113 + Math.round(19 * t)
-      return `rgb(${r}, ${g}, ${b})`
-    }
-  }
-
-  const getTextColor = (score: number): string => {
-    if (score === 0) return 'rgb(107, 114, 128)'
-    if (score >= 60) return 'rgb(17, 24, 39)'
-    return 'rgb(255, 255, 255)'
-  }
-
   const handleStudentClick = useCallback((studentName: string) => {
     const info = studentDataMap.get(studentName)
     if (info) {
@@ -552,6 +498,12 @@ const CourseDetail = () => {
     return (
       <div className="bg-red-50 border border-red-200 rounded-lg p-6">
         <div className="text-red-800">Error: {error instanceof Error ? error.message : 'An error occurred while loading course details'}</div>
+        <button
+          onClick={() => refetch()}
+          className="mt-3 px-4 py-2 bg-danger-600 text-white text-sm font-semibold rounded-lg hover:bg-danger-700 transition-colors"
+        >
+          Try Again
+        </button>
       </div>
     )
   }
@@ -565,102 +517,20 @@ const CourseDetail = () => {
   }
 
   const avgScore = getAverageScore()
-  const scoreTextColor = avgScore >= 70 ? 'text-emerald-700' : avgScore >= 50 ? 'text-amber-700' : avgScore === 0 ? 'text-secondary-400' : 'text-red-700'
 
   return (
     <div className="space-y-6">
-      <div className="flex items-center justify-between">
-        <div>
-          <h1 className="text-2xl font-bold text-secondary-900">
-            {data.course.code} - {data.course.name}
-          </h1>
-          <div className="flex items-center gap-2 mt-2">
-            {canEdit && (
-              <button
-                onClick={() => setIsEditModalOpen(true)}
-                className="bg-secondary-100 text-secondary-700 px-3 py-1.5 rounded-lg hover:bg-secondary-200 flex items-center space-x-1.5 transition-colors text-sm"
-              >
-                <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M16.862 4.487l1.687-1.688a1.875 1.875 0 112.652 2.652L10.582 16.07a4.5 4.5 0 01-1.897 1.13L6 18l.8-2.685a4.5 4.5 0 011.13-1.897l8.932-8.931zm0 0L19.5 7.125M18 14v4.75A2.25 2.25 0 0115.75 21H5.25A2.25 2.25 0 013 18.75V8.25A2.25 2.25 0 015.25 6H10" />
-                </svg>
-                <span>Edit</span>
-              </button>
-            )}
-            {canDelete && (
-              <button
-                onClick={() => setIsDeleteConfirmOpen(true)}
-                className="bg-danger-50 text-danger-700 px-3 py-1.5 rounded-lg hover:bg-danger-100 flex items-center space-x-1.5 transition-colors text-sm"
-              >
-                <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0" />
-                </svg>
-                <span>Delete</span>
-              </button>
-            )}
-          </div>
-        </div>
-        <div className="flex items-center gap-2">
-          <button
-            onClick={() => setIsFileUploadModalOpen(true)}
-            className="bg-primary-600 text-white px-4 py-2 rounded-lg hover:bg-primary-700 flex items-center space-x-2 transition-colors"
-          >
-            <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M7 16a4 4 0 01-.88-7.903A5 5 0 1115.9 6L16 6a5 5 0 011 9.9M15 13l-3-3m0 0l-3 3m3-3v12" />
-            </svg>
-            <span>Import File</span>
-          </button>
-        </div>
-      </div>
-
-      <div className="grid grid-cols-1 md:grid-cols-4 gap-4">
-        <Card variant="flat" className="bg-primary-50 border-primary-200">
-          <div className="flex items-center space-x-4">
-            <div className="p-3 bg-primary-100 rounded-xl">
-              <BookOpenIcon className="h-8 w-8 text-primary-700" />
-            </div>
-            <div>
-              <p className="text-sm text-secondary-600 font-medium">Credits</p>
-              <p className="text-3xl font-bold text-primary-700">{data.course.credits}</p>
-            </div>
-          </div>
-        </Card>
-
-        <Card variant="flat" className="bg-cyan-50 border-cyan-200">
-          <div className="flex items-center space-x-4">
-            <div className="p-3 bg-cyan-100 rounded-xl">
-              <UsersIcon className="h-8 w-8 text-cyan-700" />
-            </div>
-            <div className="min-w-0">
-              <p className="text-sm text-secondary-600 font-medium">Instructors</p>
-              <p className="text-lg font-bold text-cyan-700 truncate">{getInstructorNames()}</p>
-            </div>
-          </div>
-        </Card>
-
-        <Card variant="flat" className="bg-emerald-50 border-emerald-200">
-          <div className="flex items-center space-x-4">
-            <div className="p-3 bg-emerald-100 rounded-xl">
-              <AcademicCapIcon className="h-8 w-8 text-emerald-700" />
-            </div>
-            <div>
-              <p className="text-sm text-secondary-600 font-medium">Avg Score</p>
-              <p className={`text-3xl font-bold ${scoreTextColor}`}>{avgScore}</p>
-            </div>
-          </div>
-        </Card>
-
-        <Card variant="flat" className="bg-violet-50 border-violet-200">
-          <div className="flex items-center space-x-4">
-            <div className="p-3 bg-violet-100 rounded-xl">
-              <ChartBarIcon className="h-8 w-8 text-violet-700" />
-            </div>
-            <div>
-              <p className="text-sm text-secondary-600 font-medium">Learning Outcomes</p>
-              <p className="text-3xl font-bold text-violet-700">{data.learningOutcomes?.length || 0}</p>
-            </div>
-          </div>
-        </Card>
-      </div>
+      <CourseHeader
+        course={data.course}
+        avgScore={avgScore}
+        loCount={data.learningOutcomes?.length || 0}
+        canEdit={canEdit}
+        canDelete={canDelete}
+        onEdit={() => setIsEditModalOpen(true)}
+        onDelete={() => setIsDeleteConfirmOpen(true)}
+        onImport={() => setIsFileUploadModalOpen(true)}
+        getInstructorNames={getInstructorNames}
+      />
 
       <Card>
         <div className="flex justify-between items-center mb-4">
@@ -723,193 +593,22 @@ const CourseDetail = () => {
         {data.learningOutcomes && data.learningOutcomes.length > 0 ? (
           <>
             {loChartView === 'radar' && (
-              <ChartWidget
-                title=""
-                type="radar"
-                series={[{
-                  name: 'Average Score',
-                  data: data.learningOutcomes.map(lo => Math.round(getLOPerformance(lo.code) * 10) / 10)
-                }]}
-                options={{
-                  xaxis: {
-                    categories: data.learningOutcomes.map(lo => lo.code)
-                  },
-                  yaxis: {
-                    show: false,
-                    min: 0,
-                    max: 100
-                  },
-                  fill: {
-                    opacity: 0.3,
-                    colors: ['#6366f1']
-                  },
-                  stroke: {
-                    colors: ['#6366f1']
-                  },
-                  colors: ['#6366f1'],
-                  markers: {
-                    size: 4
-                  },
-                  dataLabels: {
-                    enabled: true,
-                    background: {
-                      enabled: true,
-                      borderRadius: 2,
-                    }
-                  },
-                  plotOptions: {
-                    radar: {
-                      polygons: {
-                        strokeColors: '#e5e7eb',
-                        connectorColors: '#e5e7eb',
-                      }
-                    }
-                  }
-                }}
-                height={320}
-                className="shadow-none border-0 p-0 [&>div]:p-0"
+              <LoRadarChart
+                learningOutcomes={data.learningOutcomes}
+                loScores={data.loScores}
               />
             )}
-            {loChartView === 'boxplot' && (
-              <div className="space-y-6">
-                {boxPlotData.length > 0 ? (
-                  boxPlotData.map((box) => {
-                    const W = 500
-                    const PAD = 12
-                    const scale = W / 100
-                    const x = (v: number) => PAD + v * scale
-
-                    const getBoxColor = (median: number) => {
-                      if (median >= 80) return { stroke: '#15803d', fill: '#bbf7d0', whisker: '#4ade80' }
-                      if (median >= 60) return { stroke: '#a16207', fill: '#fef08a', whisker: '#facc15' }
-                      return { stroke: '#b91c1c', fill: '#fecaca', whisker: '#f87171' }
-                    }
-                    const c = getBoxColor(box.median)
-
-                    return (
-                      <div key={box.code}>
-                        <div className="flex items-center justify-between mb-1.5">
-                          <span className="text-sm font-semibold text-secondary-800">{box.code}</span>
-                          <div className="flex items-center gap-3 text-xs text-secondary-500 tabular-nums">
-                            <span>Min <span className="font-semibold text-secondary-700">{box.min}</span></span>
-                            <span>Q1 <span className="font-semibold text-secondary-700">{box.q1}</span></span>
-                            <span>Med <span className="font-semibold text-secondary-700">{box.median}</span></span>
-                            <span>Q3 <span className="font-semibold text-secondary-700">{box.q3}</span></span>
-                            <span>Max <span className="font-semibold text-secondary-700">{box.max}</span></span>
-                          </div>
-                        </div>
-                        <svg width="100%" viewBox={`0 0 ${W + PAD * 2} 44`} preserveAspectRatio="none" className="overflow-visible">
-                          <rect x={PAD} y={19} width={W} height={2} rx={1} fill="#f3f4f6" />
-                          <line x1={x(box.min)} y1={20} x2={x(box.q1)} y2={20} stroke="#9ca3af" strokeWidth={1.5} />
-                          <line x1={x(box.q3)} y1={20} x2={x(box.max)} y2={20} stroke="#9ca3af" strokeWidth={1.5} />
-                          <line x1={x(box.min)} y1={12} x2={x(box.min)} y2={28} stroke="#6b7280" strokeWidth={2} strokeLinecap="round" />
-                          <line x1={x(box.max)} y1={12} x2={x(box.max)} y2={28} stroke="#6b7280" strokeWidth={2} strokeLinecap="round" />
-                          <rect
-                            x={x(box.q1)} y={8} width={x(box.q3) - x(box.q1)} height={24}
-                            fill={c.fill}
-                            stroke={c.stroke}
-                            strokeWidth={2}
-                            rx={4}
-                          />
-                          <line
-                            x1={x(box.median)} y1={6} x2={x(box.median)} y2={34}
-                            stroke={c.stroke} strokeWidth={3} strokeLinecap="round"
-                          />
-                          <polygon
-                            points={`${x(box.mean)},${14} ${x(box.mean) + 5},${20} ${x(box.mean)},${26} ${x(box.mean) - 5},${20}`}
-                            fill="#4f46e5"
-                            stroke="#fff"
-                            strokeWidth={1.5}
-                          />
-                          {[0, 25, 50, 75, 100].map((tick) => (
-                            <text key={tick} x={x(tick)} y={42} fontSize={10} textAnchor="middle" fill="#9ca3af" fontFamily="ui-monospace, monospace">
-                              {tick}
-                            </text>
-                          ))}
-                        </svg>
-                      </div>
-                    )
-                  })
-                ) : (
-                  <p className="text-secondary-500 text-center py-4">No data available</p>
-                )}
-              </div>
-            )}
+            {loChartView === 'boxplot' && <BoxPlotChart data={boxPlotData} />}
             {loChartView === 'heatmap' && (
-              <div className="overflow-x-auto">
-                <table className="w-full border-collapse text-sm">
-                  <thead>
-                    <tr className="bg-secondary-100">
-                      <th className="sticky left-0 bg-secondary-100 px-2 py-1.5 text-left text-xs font-semibold text-secondary-700 border-b border-r border-secondary-200 z-10 min-w-[120px] max-w-[180px]">
-                        Student
-                      </th>
-                      {heatmapData.loCodes.map((loCode) => (
-                        <th key={loCode} className="px-2 py-1.5 text-center text-xs font-semibold text-secondary-700 border-b border-secondary-200 min-w-[60px]">
-                          {loCode}
-                        </th>
-                      ))}
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {heatmapData.students.map((student, idx) => (
-                      <tr
-                        key={student.studentId}
-                        className={`cursor-pointer hover:bg-secondary-100/50 ${idx % 2 === 0 ? 'bg-white' : 'bg-secondary-50/50'}`}
-                        onClick={() => handleStudentClick(student.studentName)}
-                      >
-                        <td className="sticky left-0 px-2 py-1.5 text-sm font-medium text-secondary-900 border-b border-r border-secondary-200 truncate min-w-[120px] max-w-[180px]" style={{ backgroundColor: idx % 2 === 0 ? '#fff' : '#f9fafbfa' }}>
-                          {student.studentName}
-                        </td>
-                        {heatmapData.loCodes.map((loCode) => {
-                          const score = student.loScores[loCode] ?? 0
-                          const bgColor = getHeatmapColor(score)
-                          const textColor = getTextColor(score)
-                          return (
-                            <td
-                              key={loCode}
-                              className="px-2 py-1.5 text-center text-xs font-medium border-b border-secondary-200"
-                              style={{ backgroundColor: bgColor, color: textColor }}
-                            >
-                              {score > 0 ? score.toFixed(1) : '−'}
-                            </td>
-                          )
-                        })}
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-            )}
-            {loChartView === 'heatmap' && heatmapData.students.length > 0 && (
-              <div className="mt-4 flex items-center justify-center gap-1">
-                {[0, 25, 50, 75, 100].map((val) => (
-                  <div key={val} className="flex items-center gap-1">
-                    <div className="w-4 h-4 rounded" style={{ backgroundColor: getHeatmapColor(val) }} />
-                    <span className="text-xs text-secondary-600 tabular-nums">{val}%</span>
-                  </div>
-                ))}
-              </div>
+              <StudentHeatmap
+                loCodes={heatmapData.loCodes}
+                students={heatmapData.students}
+                onStudentClick={handleStudentClick}
+              />
             )}
             {boxPlotData.length > 0 && loChartView === 'boxplot' && (
               <div className="mt-5 pt-4 border-t border-secondary-100">
-                <div className="flex items-center justify-center gap-5 text-xs text-secondary-500">
-                  <div className="flex items-center gap-1.5">
-                    <div className="w-5 h-3 rounded border-2 border-emerald-700 bg-emerald-200" />
-                    <span>IQR (Q1–Q3)</span>
-                  </div>
-                  <div className="flex items-center gap-1.5">
-                    <div className="w-0.5 h-4 bg-secondary-800 rounded-full" />
-                    <span>Median</span>
-                  </div>
-                  <div className="flex items-center gap-1.5">
-                    <div className="w-2.5 h-2.5 bg-indigo-600 rotate-45 rounded-[1px]" />
-                    <span>Mean</span>
-                  </div>
-                  <div className="flex items-center gap-1.5">
-                    <div className="w-0.5 h-3 bg-gray-400 rounded-full mx-1" />
-                    <span>Whiskers (Min–Max)</span>
-                  </div>
-                </div>
+                <BoxPlotLegend />
               </div>
             )}
             <div className="mt-6 grid grid-cols-1 md:grid-cols-2 gap-3">
@@ -1055,147 +754,18 @@ const CourseDetail = () => {
                 className="shadow-none border-0 p-0 [&>div]:p-0"
               />
             )}
-            {assessmentChartView === 'boxplot' && (
-              <div className="space-y-6">
-                {assessmentBoxPlotData.length > 0 ? (
-                  assessmentBoxPlotData.map((box) => {
-                    const W = 500
-                    const PAD = 12
-                    const scale = W / 100
-                    const x = (v: number) => PAD + v * scale
-
-                    const getBoxColor = (median: number) => {
-                      if (median >= 80) return { stroke: '#15803d', fill: '#bbf7d0', whisker: '#4ade80' }
-                      if (median >= 60) return { stroke: '#a16207', fill: '#fef08a', whisker: '#facc15' }
-                      return { stroke: '#b91c1c', fill: '#fecaca', whisker: '#f87171' }
-                    }
-                    const c = getBoxColor(box.median)
-
-                    return (
-                      <div key={box.code}>
-                        <div className="flex items-center justify-between mb-1.5">
-                          <span className="text-sm font-semibold text-secondary-800">{box.code}</span>
-                          <div className="flex items-center gap-3 text-xs text-secondary-500 tabular-nums">
-                            <span>Min <span className="font-semibold text-secondary-700">{box.min}</span></span>
-                            <span>Q1 <span className="font-semibold text-secondary-700">{box.q1}</span></span>
-                            <span>Med <span className="font-semibold text-secondary-700">{box.median}</span></span>
-                            <span>Q3 <span className="font-semibold text-secondary-700">{box.q3}</span></span>
-                            <span>Max <span className="font-semibold text-secondary-700">{box.max}</span></span>
-                          </div>
-                        </div>
-                        <svg width="100%" viewBox={`0 0 ${W + PAD * 2} 44`} preserveAspectRatio="none" className="overflow-visible">
-                          <rect x={PAD} y={19} width={W} height={2} rx={1} fill="#f3f4f6" />
-                          <line x1={x(box.min)} y1={20} x2={x(box.q1)} y2={20} stroke="#9ca3af" strokeWidth={1.5} />
-                          <line x1={x(box.q3)} y1={20} x2={x(box.max)} y2={20} stroke="#9ca3af" strokeWidth={1.5} />
-                          <line x1={x(box.min)} y1={12} x2={x(box.min)} y2={28} stroke="#6b7280" strokeWidth={2} strokeLinecap="round" />
-                          <line x1={x(box.max)} y1={12} x2={x(box.max)} y2={28} stroke="#6b7280" strokeWidth={2} strokeLinecap="round" />
-                          <rect
-                            x={x(box.q1)} y={8} width={x(box.q3) - x(box.q1)} height={24}
-                            fill={c.fill}
-                            stroke={c.stroke}
-                            strokeWidth={2}
-                            rx={4}
-                          />
-                          <line
-                            x1={x(box.median)} y1={6} x2={x(box.median)} y2={34}
-                            stroke={c.stroke} strokeWidth={3} strokeLinecap="round"
-                          />
-                          <polygon
-                            points={`${x(box.mean)},${14} ${x(box.mean) + 5},${20} ${x(box.mean)},${26} ${x(box.mean) - 5},${20}`}
-                            fill="#4f46e5"
-                            stroke="#fff"
-                            strokeWidth={1.5}
-                          />
-                          {[0, 25, 50, 75, 100].map((tick) => (
-                            <text key={tick} x={x(tick)} y={42} fontSize={10} textAnchor="middle" fill="#9ca3af" fontFamily="ui-monospace, monospace">
-                              {tick}
-                            </text>
-                          ))}
-                        </svg>
-                      </div>
-                    )
-                  })
-                ) : (
-                  <p className="text-secondary-500 text-center py-4">No data available</p>
-                )}
-              </div>
-            )}
+            {assessmentChartView === 'boxplot' && <BoxPlotChart data={assessmentBoxPlotData} />}
             {assessmentChartView === 'boxplot' && assessmentBoxPlotData.length > 0 && (
               <div className="mt-5 pt-4 border-t border-secondary-100">
-                <div className="flex items-center justify-center gap-5 text-xs text-secondary-500">
-                  <div className="flex items-center gap-1.5">
-                    <div className="w-5 h-3 rounded border-2 border-emerald-700 bg-emerald-200" />
-                    <span>IQR (Q1–Q3)</span>
-                  </div>
-                  <div className="flex items-center gap-1.5">
-                    <div className="w-0.5 h-4 bg-secondary-800 rounded-full" />
-                    <span>Median</span>
-                  </div>
-                  <div className="flex items-center gap-1.5">
-                    <div className="w-2.5 h-2.5 bg-indigo-600 rotate-45 rounded-[1px]" />
-                    <span>Mean</span>
-                  </div>
-                  <div className="flex items-center gap-1.5">
-                    <div className="w-0.5 h-3 bg-gray-400 rounded-full mx-1" />
-                    <span>Whiskers (Min–Max)</span>
-                  </div>
-                </div>
+                <BoxPlotLegend />
               </div>
             )}
             {assessmentChartView === 'heatmap' && (
-              <div className="overflow-x-auto">
-                <table className="w-full border-collapse text-sm">
-                  <thead>
-                    <tr className="bg-secondary-100">
-                      <th className="sticky left-0 bg-secondary-100 px-2 py-1.5 text-left text-xs font-semibold text-secondary-700 border-b border-r border-secondary-200 z-10 min-w-[120px] max-w-[180px]">
-                        Student
-                      </th>
-                      {assessmentHeatmap.assessments.map((a) => (
-                        <th key={a.id} className="px-2 py-1.5 text-center text-xs font-semibold text-secondary-700 border-b border-secondary-200 min-w-[60px]">
-                          <div className="truncate max-w-[80px]" title={a.name}>{a.name}</div>
-                        </th>
-                      ))}
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {assessmentHeatmap.students.map((student, idx) => (
-                      <tr
-                        key={idx}
-                        className={`cursor-pointer hover:bg-secondary-100/50 ${idx % 2 === 0 ? 'bg-white' : 'bg-secondary-50/50'}`}
-                        onClick={() => handleStudentClick(student.name)}
-                      >
-                        <td className="sticky left-0 px-2 py-1.5 text-sm font-medium text-secondary-900 border-b border-r border-secondary-200 truncate min-w-[120px] max-w-[180px]" style={{ backgroundColor: idx % 2 === 0 ? '#fff' : '#f9fafbfa' }}>
-                          {student.name}
-                        </td>
-                        {assessmentHeatmap.assessments.map((a) => {
-                          const pct = student.scores[a.id] ?? 0
-                          const bgColor = getHeatmapColor(pct)
-                          const textColor = getTextColor(pct)
-                          return (
-                            <td
-                              key={a.id}
-                              className="px-2 py-1.5 text-center text-xs font-medium border-b border-secondary-200"
-                              style={{ backgroundColor: bgColor, color: textColor }}
-                            >
-                              {pct > 0 ? pct.toFixed(1) : '−'}
-                            </td>
-                          )
-                        })}
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-            )}
-            {assessmentChartView === 'heatmap' && assessmentHeatmap.students.length > 0 && (
-              <div className="mt-4 flex items-center justify-center gap-1">
-                {[0, 25, 50, 75, 100].map((val) => (
-                  <div key={val} className="flex items-center gap-1">
-                    <div className="w-4 h-4 rounded" style={{ backgroundColor: getHeatmapColor(val) }} />
-                    <span className="text-xs text-secondary-600 tabular-nums">{val}%</span>
-                  </div>
-                ))}
-              </div>
+              <AssessmentHeatmap
+                assessments={assessmentHeatmap.assessments}
+                students={assessmentHeatmap.students}
+                onStudentClick={handleStudentClick}
+              />
             )}
 </>
             )}
